+++ backend/flowmeet-api/src/main/java/kr/flowmeet/api/project/facade/ProjectUrlFacade.java
@@ -7,7 +7,6 @@ import kr.flowmeet.api.common.exception.ApiException;
 import kr.flowmeet.api.project.dto.request.ProjectUrlRequest;
 import kr.flowmeet.api.project.dto.response.ProjectUrlResponse;
 import kr.flowmeet.domain.project.entity.ProjectMember;
-import kr.flowmeet.domain.project.entity.ProjectMemberRole;
 import kr.flowmeet.domain.project.entity.ProjectUrl;
 import kr.flowmeet.domain.project.exception.ProjectErrorCode;
 import kr.flowmeet.domain.project.service.ProjectMemberService;
@@ -24,7 +23,9 @@ public class ProjectUrlFacade {
     @Transactional
     public ProjectUrlResponse addUrl(final Long userId, final Long projectId, final ProjectUrlRequest request) {
         ProjectMember requesterMember = projectMemberService.findByProjectIdAndUserId(projectId, userId);
-        validateMemberCanEdit(requesterMember);
+        if (!requesterMember.canEdit()) {
+            throw new ApiException(ProjectErrorCode.PROJECT_ACCESS_DENIED);
+        }
 
         ProjectUrl projectUrl = projectUrlService.create(
                 ProjectUrl.builder()
@@ -40,7 +41,9 @@ public class ProjectUrlFacade {
     public ProjectUrlResponse updateUrl(final Long userId, final Long projectId, final Long urlId,
                                         final ProjectUrlRequest request) {
         ProjectMember requesterMember = projectMemberService.findByProjectIdAndUserId(projectId, userId);
-        validateMemberCanEdit(requesterMember);
+        if (!requesterMember.canEdit()) {
+            throw new ApiException(ProjectErrorCode.PROJECT_ACCESS_DENIED);
+        }
 
         ProjectUrl projectUrl = projectUrlService.findByIdAndProjectId(urlId, projectId);
         projectUrl.updateUrl(request.url());
@@ -51,15 +54,12 @@ public class ProjectUrlFacade {
     @Transactional
     public void deleteUrl(final Long userId, final Long projectId, final Long urlId) {
         ProjectMember requesterMember = projectMemberService.findByProjectIdAndUserId(projectId, userId);
-        validateMemberCanEdit(requesterMember);
+        if (!requesterMember.canEdit()) {
+            throw new ApiException(ProjectErrorCode.PROJECT_ACCESS_DENIED);
+        }
 
         ProjectUrl projectUrl = projectUrlService.findByIdAndProjectId(urlId, projectId);
         projectUrlService.delete(projectUrl);
     }
 
-    private void validateMemberCanEdit(final ProjectMember member) {
-        if (member.getRole() == ProjectMemberRole.VIEWER) {
-            throw new ApiException(ProjectErrorCode.PROJECT_ACCESS_DENIED);
-        }
-    }
 }
